bulk: restructure crate, add pos_to_point
